requirements.txt
streamlit==1.40.1
requests==2.32.3
python-dotenv==1.0.1
openai==1.55.0
